list of files to filter are now on their own
